角色改为单角色
