--- abeille-basic-hypervisor/src/main/java/top/abeille/basic/hypervisor/entity/UserRole.java
@@ -1,98 +1,0 @@
-/*
- * Copyright (c) 2019. Abeille All Right Reserved.
- */
-package top.abeille.basic.hypervisor.entity;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-/**
- * Model class for UserRole
- *
- * @author liwenqiang 2018/12/4 10:09
- **/
-@Entity
-@Table(name = "user_role")
-public class UserRole {
-
-    /**
-     * 主键
-     */
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    /**
-     * 用户主键
-     */
-    @Column(name = "user_id")
-    private Long userId;
-    /**
-     * 角色主键
-     */
-    @Column(name = "role_id")
-    private Long roleId;
-
-    /**
-     * 是否有效
-     */
-    @Column(name = "is_enabled")
-    private Boolean enabled;
-    /**
-     * 修改人
-     */
-    @Column(name = "modifier")
-    private Long modifier;
-    /**
-     * 修改时间
-     */
-    @Column(name = "modify_time")
-    private LocalDateTime modifyTime;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public Long getRoleId() {
-        return roleId;
-    }
-
-    public void setRoleId(Long roleId) {
-        this.roleId = roleId;
-    }
-
-    public Boolean getEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(Boolean enabled) {
-        this.enabled = enabled;
-    }
-
-    public Long getModifier() {
-        return modifier;
-    }
-
-    public void setModifier(Long modifier) {
-        this.modifier = modifier;
-    }
-
-    public LocalDateTime getModifyTime() {
-        return modifyTime;
-    }
-
-    public void setModifyTime(LocalDateTime modifyTime) {
-        this.modifyTime = modifyTime;
-    }
-}
